requirements.txt
anyio==4.7.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
psycopg==3.2.3
python-dotenv==1.0.1
python-telegram-bot==21.8
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
